Use grid format by default
